requirements.txt
discord.py~=2.6.3
Flask~=3.1.2